Revert "[Functional] Migrate backend(Base database) to Firebase Realtime Database"
This reverts commit 5e34ee1c

+++ app/src/main/java/com/example/shereats/utils/ImageUtil.kt
@@ -154,7 +154,7 @@ class ImageUtil {
         /**
          * Return the badge color according to its rarity
          */
-        fun getBadgeColor(rarity: Long): Int{
+        fun getBadgeColor(rarity: Int): Int{
             return ConstantUtil.MAP_BADGE_COLOR[rarity] ?: Color.parseColor("#000000")
         }
